move func to file that uses it
